EasyMDE actions
